network module moved to app | server url is fetched from build config
