+++ app/build.gradle
@@ -23,6 +23,11 @@ android {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            buildConfigField "String", "SERVER_URL", "\"https://api.merakilearn.org/\""
+        }
+
+        debug {
+            buildConfigField "String", "SERVER_URL", "\"https://dev-api.merakilearn.org/apiDocs/\""
         }
     }
     compileOptions {
