+++ app/src/main/java/org/merakilearn/di/Modules.kt
@@ -1,9 +1,16 @@
 package org.merakilearn.di
 
 import android.app.Application
+import com.google.gson.Gson
+import com.google.gson.GsonBuilder
+import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
+import okhttp3.Interceptor
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
 import org.koin.android.ext.koin.androidApplication
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
+import org.merakilearn.BuildConfig
 import org.merakilearn.core.navigator.AppModuleNavigator
 import org.merakilearn.datasource.ApplicationRepo
 import org.merakilearn.datasource.network.SaralApi
@@ -12,9 +19,12 @@ import org.merakilearn.ui.home.HomeViewModel
 import org.merakilearn.ui.more.MoreViewModel
 import org.merakilearn.ui.onboarding.LoginViewModel
 import org.merakilearn.ui.onboarding.WelcomeViewModel
-import org.navgurukul.learn.courses.db.CoursesDatabase
 import org.navgurukul.chat.core.repo.AuthenticationRepository
+import org.navgurukul.learn.courses.db.CoursesDatabase
 import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+
 
 val viewModelModule = module {
     viewModel { LoginViewModel(get()) }
@@ -34,6 +44,47 @@ val apiModule = module {
     single { provideUserApi(get()) }
 }
 
+
+val networkModule = module {
+
+    fun provideHttpClient(): OkHttpClient {
+        val okHttpClientBuilder = OkHttpClient.Builder()
+        okHttpClientBuilder.addInterceptor(provideLogInterceptor())
+        okHttpClientBuilder.connectTimeout(5, TimeUnit.MINUTES)
+        okHttpClientBuilder.readTimeout(5, TimeUnit.MINUTES)
+
+        return okHttpClientBuilder.build()
+    }
+
+    fun provideGson(): Gson {
+        return GsonBuilder().setLenient().create()
+    }
+
+
+    fun provideRetrofit(factory: Gson, client: OkHttpClient): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(BuildConfig.SERVER_URL)
+            .addConverterFactory(GsonConverterFactory.create(factory))
+            .addCallAdapterFactory(CoroutineCallAdapterFactory())
+            .client(client)
+            .build()
+    }
+
+    single { provideHttpClient() }
+    single { provideGson() }
+    single { provideRetrofit(get(), get()) }
+
+}
+
+fun provideLogInterceptor(): Interceptor {
+    val logging = HttpLoggingInterceptor()
+    logging.level = if (BuildConfig.DEBUG)
+        HttpLoggingInterceptor.Level.BODY
+    else
+        HttpLoggingInterceptor.Level.NONE
+    return logging
+}
+
 val repositoryModule = module {
     fun provideAppRepo(
         api: SaralApi,
@@ -51,4 +102,4 @@ val repositoryModule = module {
 
     single { provideAppRepo(get(), androidApplication(),get(), get()) }
 }
-val appModules = arrayListOf(viewModelModule, apiModule, factoryModule, repositoryModule)
+val appModules = arrayListOf(viewModelModule, apiModule, networkModule, factoryModule, repositoryModule)
